9624: Move Filters utils under Filters/
