+++ app/javascript/components/QuestionFilter/component.js
@@ -5,7 +5,7 @@ import Popover from 'react-bootstrap/Popover';
 import OverlayTrigger from 'react-bootstrap/OverlayTrigger';
 import { inject, observer } from 'mobx-react';
 
-import { getButtonHintString, getItemNameFromId } from '../search/utils';
+import { getButtonHintString, getItemNameFromId } from '../Filters/utils';
 import ConditionSetFormField from '../ConditionSetFormField/component';
 
 @inject('filtersStore')
